Use float division from __future__

+++ stingray/utils.py
@@ -1,3 +1,4 @@
+from __future__ import division
 import numpy as np
 import warnings
 
@@ -63,7 +64,7 @@ def rebin_data(x, y, dx_new, method='sum'):
     assert dx_new >= dx_old, "New frequency resolution must be larger than " \
                              "old frequency resolution."
 
-    step_size = np.float(dx_new)/np.float(dx_old)
+    step_size = dx_new/dx_old
 
     output = []
     for i in np.arange(0, y.shape[0], step_size):
@@ -95,7 +96,7 @@ def rebin_data(x, y, dx_new, method='sum'):
 
     tseg = x[-1]-x[0]+dx_old
 
-    if tseg/dx_new % 1.0 > 0.0:
+    if tseg/dx_new % 1 > 0:
         ybin = ybin[:-1]
 
     xbin = np.arange(ybin.shape[0])*dx_new + x[0]-dx_old/2.+dx_new/2.
